requirements.txt
# Core dependencies for Agent Platform
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
python-consul==1.1.0

# Model Context Protocol
mcp>=1.0.0

# Additional utilities
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1

